requirements.txt
openai == 1.9.0
antlr4-python3-runtime==4.13.1
anthropic == 0.20.0